+++ src/aggregateFunction.ts
@@ -30,6 +30,46 @@ export class AggregateFunction {
     return new Condition(new Expression(this), ComparisonOperator.Equal, new Expression(new Binder(value)))
   }
 
+  public ne(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.NotEqual, new Expression(value))
+  }
+
+  public ne$(value:number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.NotEqual, new Expression(new Binder(value)))
+  }
+
+  public gt(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.GreaterThan, new Expression(value))
+  }
+
+  public gt$(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.GreaterThan, new Expression(new Binder(value)))
+  }
+
+  public ge(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.GreaterOrEqual, new Expression(value))
+  }
+
+  public ge$(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.GreaterOrEqual, new Expression(new Binder(value)))
+  }
+
+  public lt(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.LesserThan, new Expression(value))
+  }
+
+  public lt$(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.LesserThan, new Expression(new Binder(value)))
+  }
+
+  public le(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.LesserOrEqual, new Expression(value))
+  }
+
+  public le$(value: number): Condition {
+    return new Condition(new Expression(this), ComparisonOperator.LesserOrEqual, new Expression(new Binder(value)))
+  }
+
   public getStmt(data: BuilderData): string {
     return `${this.funcName}(${this.expression.getStmt(data, { withOuterBracket: false })})`
   }