+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Move Expression into Expression.ts file
 - Move Condition into Condition.ts file
 - Move Operand class into Operand.ts file
+- Rename models to types and move it inside models folder
 
 ### Added
 - Having step can contain aggregate function condition